fix load order
